Remove python evaluator
